requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
numpy==1.24.3
tensorflow-cpu==2.17.0  # Use CPU version to avoid CUDA issues
keras==3.5.0  # Ensure compatibility with TensorFlow 2.17
Pillow==10.4.0  # For image processing
gunicorn==23.0.0  # For deploying the Flask app in production
